Split gateway pool edit component
